+++ pkg/errors/details.go
@@ -30,20 +30,30 @@ func (e *Error) addDetails(details ...proto.Message) {
 
 // WithDetails returns the error with the given details set.
 // This appends to any existing details in the Error.
-func (e Error) WithDetails(details ...proto.Message) Error {
-	e.addDetails(details...)
-	return e
+func (e *Error) WithDetails(details ...proto.Message) *Error {
+	if e == nil {
+		return e
+	}
+	dup := *e
+	dup.addDetails(details...)
+	return &dup
 }
 
 // WithDetails returns a new error from the definition, and sets the given details.
-func (d Definition) WithDetails(details ...proto.Message) Error {
+func (d *Definition) WithDetails(details ...proto.Message) *Error {
+	if d == nil {
+		return nil
+	}
 	e := build(d, 0) // Don't refactor this to build(...).WithDetails(...)
 	e.addDetails(details...)
 	return e
 }
 
 // Details of the error. Usually structs from ttnpb or google.golang.org/genproto/googleapis/rpc/errdetails.
-func (e Error) Details() (details []proto.Message) {
+func (e *Error) Details() (details []proto.Message) {
+	if e == nil {
+		return nil
+	}
 	if e.cause != nil {
 		details = append(details, Details(e.cause)...)
 	}
@@ -54,7 +64,7 @@ func (e Error) Details() (details []proto.Message) {
 }
 
 // Details are not present in the error definition, so this just returns nil.
-func (d Definition) Details() []proto.Message { return nil }
+func (*Definition) Details() []proto.Message { return nil }
 
 // Details gets the details of the error.
 func Details(err error) []proto.Message {