+++ pkg/errors/errors.go
@@ -26,8 +26,8 @@ import (
 
 // New returns an error that formats as the given text.
 // This way of creating errors should be avoided if possible.
-func New(text string) Error {
-	return build(Definition{
+func New(text string) *Error {
+	return build(&Definition{
 		namespace:     namespace(2),
 		messageFormat: text,
 		code:          uint32(codes.Unknown),
@@ -36,7 +36,7 @@ func New(text string) Error {
 
 // Error is a rich error implementation.
 type Error struct {
-	Definition
+	*Definition
 	*stack
 	correlationID string
 	cause         error
@@ -45,15 +45,26 @@ type Error struct {
 	grpcStatus    *atomic.Value
 }
 
-func (e Error) String() string {
+func (e *Error) String() string {
+	if e == nil {
+		return ""
+	}
 	return fmt.Sprintf("error:%s (%s)", e.FullName(), e.FormatMessage(e.PublicAttributes()))
 }
 
 // Error implements the error interface.
-func (e Error) Error() string { return e.String() }
+func (e *Error) Error() string {
+	if e == nil {
+		return ""
+	}
+	return e.String()
+}
 
 // Fields implements the log.Fielder interface.
-func (e Error) Fields() map[string]interface{} {
+func (e *Error) Fields() map[string]interface{} {
+	if e == nil {
+		return nil
+	}
 	res := make(map[string]interface{})
 	pref := "error_cause"
 	for cause := Cause(e); cause != nil; cause = Cause(cause) {
@@ -75,10 +86,10 @@ type Interface interface {
 }
 
 // build an error from the definition, skipping the first frames of the call stack.
-func build(d Definition, skip int) Error {
+func build(d *Definition, skip int) *Error {
 	e := Error{Definition: d, grpcStatus: new(atomic.Value)}
 	if skip > 0 {
 		e.stack = callers(skip)
 	}
-	return e
+	return &e
 }